LayoutTests/fast/multicol/newmulticol/balance-maxheight1.html

<!DOCTYPE html>
<html>
  <head>
    <title>balancing multicol with max-height</title>
    <script src="../../../resources/js-test-pre.js"></script>
    <script>
      if (window.testRunner)
        testRunner.dumpAsText();
    </script>
  </head>
  <body>
    <p>There should be two identical boxes below.</p>
    <div id="mc" style="-webkit-columns:3; columns:3; max-height:2em; orphans:1; widows:1; color:olive; background:olive;">
      line<br>
      line<br>
      line<br>
    </div>
    <div id="ref" style="margin-top:1em; color:olive; background:olive;">
      line<br>
    </div>
    <script>
      shouldBe("mc.offsetHeight", "ref.offsetHeight");
      shouldBe("document.documentElement.scrollWidth", "document.documentElement.clientWidth");
    </script>
    <script src="../../../resources/js-test-post.js"></script>
  </body>
</html>
