third_party/blink/web_tests/fast/multicol/newmulticol/balance-maxheight1-expected.txt

PASS mc.offsetHeight is ref.offsetHeight
PASS document.documentElement.scrollWidth is document.documentElement.clientWidth
PASS successfullyParsed is true

TEST COMPLETE


There should be two identical boxes below.

line
line
line

line
